requirements.txt
requests==2.28.2
kafka-python==2.0.2
pyspark==3.1.2
pandas==1.5.3
psycopg2-binary==2.9.6